+++ src/main/java/com/petbackend/thbao/services/impl/PetReportService.java
@@ -2,6 +2,7 @@ package com.petbackend.thbao.services.impl;
 
 import com.petbackend.thbao.dtos.PetReportDTO;
 import com.petbackend.thbao.dtos.ReportImageDTO;
+import com.petbackend.thbao.exceptions.AccessDeniedException;
 import com.petbackend.thbao.exceptions.DataNotFoundException;
 import com.petbackend.thbao.models.PetMissing;
 import com.petbackend.thbao.models.PetReport;
@@ -16,6 +17,8 @@ import com.petbackend.thbao.services.IPetReportService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -37,6 +40,11 @@ public class PetReportService implements IPetReportService {
         if(!petMissing.isPresent()){
             throw new DataNotFoundException("Cannot found Pet missing");
         }
+        String name = SecurityContextHolder.getContext().getAuthentication().getName();
+        // Chỉ tạo được đơn báo cáo cho bản thân
+        if(!user.get().getPhoneNumber().equals(name)){
+            throw new AccessDeniedException("You cannot create an report for someone else");
+        }
         PetReport petReport = PetReport.builder()
                 .phoneNumber(petReportDTO.getPhoneNumber())
                 .addressFound(petReportDTO.getAddressFound())
@@ -53,20 +61,47 @@ public class PetReportService implements IPetReportService {
     }
 
     @Override
+    @PreAuthorize("hasRole('ADMIN')")
     public Page<PetReportResponse> getAllPetReport(PageRequest pageRequest) {
         return petReportRepository.findAll(pageRequest).map(PetReportResponse::fromPetReportResponse);
     }
 
     @Override
     public PetReport getPetReportById(Long id) throws DataNotFoundException {
+        String role = String.valueOf(SecurityContextHolder.getContext().getAuthentication().getAuthorities());
+        String phoneNumber = SecurityContextHolder.getContext().getAuthentication().getName();
+        User existUser = userRepository.findByPhoneNumber(phoneNumber).orElseThrow(() ->
+                new DataNotFoundException("User not exist"));
+        PetReport petReport = petReportRepository.findById(id).orElseThrow(() ->
+                new DataNotFoundException("Pet report not exist"));
+        PetMissing petMissing = petMissingRepository.findById(petReport.getPetMissing().getId()).orElseThrow(() ->
+                new DataNotFoundException("Pet missing not exist"));
+        // User bi mat thu cung (missing)/
+        User userByMissing = userRepository.findById(petMissing.getUser().getId()).orElseThrow(() ->
+                new DataNotFoundException("User not exist"));
+        // User dang bai report
+        User userByReport = userRepository.findById(petReport.getUser().getId()).orElseThrow(() ->
+                new DataNotFoundException("User not exist"));
+        if (!role.contains("ROLE_ADMIN") && !userByMissing.getPhoneNumber().equals(phoneNumber) &&
+                !userByReport.getPhoneNumber().equals(phoneNumber)){
+            throw new AccessDeniedException("Denied this report");
+        }
         return petReportRepository.findById(id).orElseThrow(()->
                 new DataNotFoundException("Cannot found Pet report"));
     }
 
     @Override
     public List<PetReport> getPetReportByUserId(Long userId) throws DataNotFoundException {
-        User existsUser = userRepository.findById(userId).orElseThrow(()->
+        User user = userRepository.findById(userId).orElseThrow(()->
                 new DataNotFoundException("Cannot found user"));
+        String role = String.valueOf(SecurityContextHolder.getContext().getAuthentication().getAuthorities());
+        String phoneNumber = SecurityContextHolder.getContext().getAuthentication().getName();
+        User existUser = userRepository.findByPhoneNumber(phoneNumber).orElseThrow(() ->
+                new DataNotFoundException("User not exist"));
+        if (userId != existUser.getId() && !role.contains("ROLE_ADMIN")){
+            // Nếu role user lấy bao cao từ id người khác thì bị lỗi
+            throw new AccessDeniedException("You cannot access orders from other people");
+        }
         List<PetReport> list = petReportRepository.findByUserId(userId);
         return list;
     }
@@ -75,8 +110,14 @@ public class PetReportService implements IPetReportService {
     public List<PetReport> getPetReportByPetMissingId(Long missingId) throws DataNotFoundException {
         PetMissing petMissing = petMissingRepository.findById(missingId).orElseThrow(()->
                 new DataNotFoundException("Cannot found Pet missing"));
-        List<PetReport> list = petReportRepository.findByPetMissingId(missingId);
-        return list;
+        User userByMissing = userRepository.findById(petMissing.getUser().getId()).orElseThrow(()->
+                new DataNotFoundException("User not exist"));
+        String role = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
+        String phoneNumber = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!role.contains("ROLE_ADMIN") && !userByMissing.getPhoneNumber().equals(phoneNumber)){
+            throw new AccessDeniedException("You denied this report");
+        }
+        return petReportRepository.findByPetMissingId(missingId);
     }
 
     @Override
@@ -92,6 +133,11 @@ public class PetReportService implements IPetReportService {
                 new DataNotFoundException("Cannot found this user"));
         PetMissing petMissing = petMissingRepository.findById(petReportDTO.getPetMissingId()).orElseThrow(()->
                 new DataNotFoundException("Cannot found this pet missing"));
+        String phoneNumber = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!petReport.getUser().getPhoneNumber().equals(phoneNumber)){
+            // chỉ update được bao cao của bản thân
+            throw new AccessDeniedException("You cannot update an missing for someone else");
+        }
         petReport.setUser(user);
         petReport.setPetMissing(petMissing);
         petReportRepository.save(petReport);
@@ -102,6 +148,13 @@ public class PetReportService implements IPetReportService {
     public void delete(Long id) throws DataNotFoundException {
         PetReport petReport = petReportRepository.findById(id).orElseThrow(()->
                 new DataNotFoundException("Cannot found this pet report"));
+        String phoneNumber = SecurityContextHolder.getContext().getAuthentication().getName();
+        String role = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
+        User user = userRepository.findByPhoneNumber(phoneNumber).orElseThrow(() ->
+                new DataNotFoundException("User not exist"));
+        if (user.getId() != petReport.getUser().getId() && !role.contains("ROLE_ADMIN")){
+            throw new AccessDeniedException("You cannot delete for someone else missing");
+        }
         petReportRepository.delete(petReport);
     }
 
@@ -109,6 +162,13 @@ public class PetReportService implements IPetReportService {
     public ReportImage createReportImage(Long petReportId, ReportImageDTO reportImageDTO) throws DataNotFoundException {
         PetReport petReport = petReportRepository.findById(petReportId).orElseThrow(()->
                 new DataNotFoundException("Cannot found this pet report"));
+        String phoneNumber = SecurityContextHolder.getContext().getAuthentication().getName();
+        String role = SecurityContextHolder.getContext().getAuthentication().getAuthorities().toString();
+        User user = userRepository.findByPhoneNumber(phoneNumber).orElseThrow(() ->
+                new DataNotFoundException("User not exist"));
+        if (user.getId() != petReport.getUser().getId() && !role.contains("ROLE_ADMIN")){
+            throw new AccessDeniedException("You cannot import image for someone else missing");
+        }
         ReportImage reportImage = ReportImage.builder()
                 .petReport(petReport)
                 .url(reportImageDTO.getUrl()).build();
